whois_child_mention_checker/requirements.txt
openai==0.28.0
pyyaml==6.0.2
telethon==1.37.0